Modifying MessagePoller connectivity architecture
MessagePoller (and decendents) will use a new connectivity architure built
an idle period specified in comm_link table.  This property will allow
connectivity to be handled in a more effecient manner and without the added
burden of specifying a connection type within the code (MnDOT didn't favor it).
Additionally, removing the cohu/pelcod-specific system attributes which are
replaced by this functionality.