allowing for numerical degrees in override field
the CAMERA_DIRECTION_OVERRIDE field may now accept integers between 0 and
360 (inclusive) in addition to the four cardinal directions. All other
values are ignored and results in the default behavior.

adjusted logic so that it only determines if an override is present once, so
that later on either the overridden value or proper tangent [can be computed
immediately and] will be returned.