temporary fix for fast failure rate
D10 noticed a slow down, and an alarmingly high failure rate count on failed
comm links. URMS/RTMS/Wizard-Property classes are not able to update the
failTime of the associated controllers, thus failure is incremented (and noted
and propagated and stored server-side in the DB) every second.

IMPORTANT: These classes probably need to be rewritten, but for now a bandaid
fix is all they get.