SiteData logic and integrity fixes
This probably relates to the CA-TRAC #579 issue.

To address the issue of an NPE occurring when trying to edit properties of
a comm_link.  Caused by a single record in the site_data table, the
site_name field was set to an empty string rather than null.

changed the logic surrounding the creation of a SiteData object to mimic
similar functionality in other parts of IRIS (a 20 second time-out).

Since the site_data table has a UNIQUE constraint on the site_name table, the
code now ensures that blank site_names ('', or just spaces) are rendered to
become null in the database. table was not given a constraint for now due to
other logic and/or existing records breaking desired constraint