change queuing of DMSXML operations
Because of the change of California's handling of DMSXML status info, the
OpQueryConfig's PhaseGetConfig is not guaranteed to succeed.  As such, the
queuing of OpQueryMsg's PhaseGetConfig then PhaseQueryMsg is subject to failing
to call the second (PhaseQueryMsg) resulting in the device remaining in a failed
state. Queuing is now handled at the poller as two separate operations, rather
than a setup of two subsequent phases in side a single operation.