proxy filter fix to prevent client crash
Filters were causing the client to crash with an index out of bounds runtime
error on the indices list on ProxyListModel. This fixes that problem, however
it is a slightly buggy fix -- causes selected items to not remain selected
or the highlighted (selected) item may clear it's selection or move to another
item in the list. items may disappear from list, that should be in it.