TransientPoller temporary fix.
Modified TransientPoller to be the same as MessagePoller for the time being.

As CalTrans is not on an all-fibre network like MnDOT, it requires "duplicate"
commands to be executed.  We think, due to the speed of fibre, that MnDOT does
not really experience much of a command queue, if any.

Should rewrite TransientPoller to exclude adding a command that is a duplicate
of the last-in command instead of any duplicate within the queue. Additionally,
contention handling within MessagePoller should also be looked at to ensure
proper order.

See TransientPoller for more information.