refactored VideoWallManager
As this feature will be required to know if a camera is in use,
VideoWallManager was refactored.  A new parent class was created to deal with
most of the camera count information, this was move from VideoWallManager into
VideoServerCoupler.  This parent class will also allow the IRIS server to pass
in a properties file (iris-server.properties) and query the video server on its
own.