new preset button alias assigment indicator
removed colors or gradients on preset button. using a simple bar instead.
this has the added benefit of being usable on a preset that has both day [home]
and night shift assigned to it.
day/home shift, bar is placed at top of button.
night shift, bar is placed at bottom of button.