added field for individual camera shift schedule
Allows for individual camera be set to a shift schedule that occurs at
specific times after the hour. If this value is defined, and set to 15,
the camera shift for this camera would occur after the actual twilight
shift, following the actual shift at 15 minutes past the hour. So if the
twilight shift occurs normally at 6:14, that camera will move one minute
later, but if the twilight shift occurs at 6:16, that camera will move at
7:15.