only shift cameras if not in use
VideoWallManager was refactored.  A new parent class was created to deal with
most of the camera count information, this was move from VideoWallManager into
VideoServerCoupler.  This parent class will also allow the IRIS server to pass
in a properties file (iris-server.properties) and query the video server on its
own.
This enables the camera shift job to not move cameras that are in use
(connected by video wall or iris client).