filtered selection issues
when filtering, a selected proxy would remain selected (for right-click
properties) if the filter was altered. will now clear any selected proxy
if filter is changed.
Found that if a filter value returned no results and then status
(another type of filtering) was filtered an IndexOutOfBounds error
occured. Altered logic to prevent this.
Also found it was necessary to clear the selection model if the status
selection was changed, as it could cause the same properties issue
above.
resolves CAIRISINT-27 and CAIRISINT-29