LCS_POLL_PERIOD_SECS w/o restart
if LCS_POLL_PERIOD_SECS changes, the new interval period will be recognized
after the completion of the next polling.
In keeping with the original intent of how LcsQueryMsgJob was setup by the
server, if the interval value is less than or equal to 5, the job will do
nothing, except repeat every 30 seconds looking to when the interval value is
valid. When the interval value is valid, it will perform a repeating job at
the specified interval.
additional job will scan at least every minute (or 30 seconds) to see if the
LCS system attribute value changed. if it does, it will remove all
LcsQueryMsgJobs and readd required jobs.